Greatly simplified sidebar code + search works now
