+++ src/feature/common/Sidebar/SidebarNode.tsx
@@ -1,13 +1,9 @@
+import { createStyles } from "@mantine/core";
 import {
-  Box,
-  ChevronIcon,
-  Collapse,
-  Group,
-  UnstyledButton,
-  createStyles,
-} from "@mantine/core";
-import { CommandManifestMap } from "@src/feature/flow_chart_panel/manifest/COMMANDS_MANIFEST";
-import { useState } from "react";
+  CommandManifestMap,
+  CommandSection,
+} from "@src/feature/flow_chart_panel/manifest/COMMANDS_MANIFEST";
+import SidebarSection from "./SidebarSection";
 
 export const useSidebarStyles = createStyles((theme) => ({
   control: {
@@ -38,38 +34,94 @@ export const useSidebarStyles = createStyles((theme) => ({
 }));
 
 type SidebarNodeProps = {
-  onClickHandle: (key: string) => void;
-  keyNode: string;
-  manifestMap: CommandManifestMap;
   depth: number;
+  manifestMap: CommandManifestMap;
+  node: CommandSection;
+  leafClickHandler: (key: string) => void;
+  query: string;
+  matchedParent: boolean;
 };
 
+const nodeTitleMatches = (query: string, node: CommandSection) =>
+  query !== "" &&
+  node.title.toLocaleLowerCase().includes(query.toLocaleLowerCase());
+
 const SidebarNode = ({
-  onClickHandle,
-  keyNode,
+  depth,
   manifestMap,
+  node,
+  leafClickHandler,
+  query,
+  matchedParent = false,
 }: SidebarNodeProps) => {
   const { classes } = useSidebarStyles();
-  const commands = manifestMap[keyNode] || [];
-  const [open, setOpen] = useState(false);
+
+  if (node.title === "ROOT") {
+    if (!node.children) return null;
+
+    return (
+      <div>
+        {node.children.map((c) =>
+          // Actually needs to be called as a function to achieve depth-first traversal,
+          // otherwise React lazily evaluates it and doesn't recurse immediately, resulting in breadth-first traversal.
+          SidebarNode({
+            node: c,
+            depth: 0,
+            manifestMap,
+            leafClickHandler,
+            query,
+            matchedParent: nodeTitleMatches(query, c),
+          })
+        )}
+      </div>
+    );
+  }
+
+  if (node.children) {
+    return (
+      <SidebarSection title={node.title} depth={depth + 1}>
+        {node.children.map((c) =>
+          SidebarNode({
+            node: c,
+            depth: depth + 1,
+            manifestMap,
+            leafClickHandler,
+            query,
+            matchedParent: nodeTitleMatches(query, c),
+          })
+        )}
+      </SidebarSection>
+    );
+  }
+
+  const key = node.key ?? "";
+
+  const commands = manifestMap[key] ?? [];
+  const lowercased = query.toLocaleLowerCase();
+  const shouldFilter = query !== "" && !matchedParent;
+  const searchMatches = shouldFilter
+    ? commands.filter(
+        (c) =>
+          c.key.toLocaleLowerCase().includes(lowercased) ||
+          c.name.toLocaleLowerCase().includes(lowercased)
+      )
+    : commands;
+  if (searchMatches.length === 0) {
+    return null;
+  }
+
   return (
-    <>
-      <Collapse in>
-        <div>
-          {commands.map((cmd) => (
-            <button
-              key={cmd.key}
-              className={classes.buttonLeafNode}
-              onClick={() => {
-                onClickHandle(cmd.key || keyNode);
-              }}
-            >
-              {cmd.key || cmd.name}
-            </button>
-          ))}
-        </div>
-      </Collapse>
-    </>
+    <SidebarSection title={node.title} depth={depth}>
+      {searchMatches.map((command) => (
+        <button
+          key={command.key}
+          className={classes.buttonLeafNode}
+          onClick={() => leafClickHandler(command.key ?? key)}
+        >
+          {command.key || command.name}
+        </button>
+      ))}
+    </SidebarSection>
   );
 };
