+++ src/feature/common/Sidebar/SidebarSection.tsx
@@ -7,6 +7,7 @@ import {
   Collapse,
 } from "@mantine/core";
 import { IconChevronLeft, IconChevronRight } from "@tabler/icons-react";
+import { Children } from "react";
 
 export const useSidebarStyles = createStyles((theme) => ({
   control: {
@@ -16,7 +17,9 @@ export const useSidebarStyles = createStyles((theme) => ({
     padding: `${theme.spacing.xs} ${theme.spacing.xs}`,
     color: "black",
     fontSize: theme.fontSizes.sm,
-    margin: "10px 20px",
+    margin: "0px 20px 5px 20px",
+
+    borderRadius: 2,
     backgroundColor: theme.colors.accent1[0],
   },
 
@@ -27,14 +30,19 @@ export const useSidebarStyles = createStyles((theme) => ({
 
 type SidebarSectionProps = {
   title: string;
-  content: React.ReactNode;
   depth: number;
+  children: React.ReactNode;
 };
 
-const SidebarSection = ({ title, content, depth }: SidebarSectionProps) => {
+const SidebarSection = ({ depth, title, children }: SidebarSectionProps) => {
   const [opened, setOpened] = useState(false);
   const { classes, theme } = useSidebarStyles();
   const ChevronIcon = theme.dir === "ltr" ? IconChevronRight : IconChevronLeft;
+
+  if (Children.toArray(children).every((child) => child === null)) {
+    return null;
+  }
+
   return (
     <>
       <UnstyledButton
@@ -60,8 +68,8 @@ const SidebarSection = ({ title, content, depth }: SidebarSectionProps) => {
       </UnstyledButton>
       <Collapse in={opened}>
         {/* padding according to the depth of the section */}
-        <div style={{ paddingLeft: `${10 + (depth + 1) * 20}px` }}>
-          {content}
+        <div style={{ paddingLeft: `${10 + (depth + 1) * 5}px` }}>
+          {children}
         </div>
       </Collapse>
     </>
